:recycle: Update files based on code review, move account, LiskService and Transactions to api/lsk folder
