+++ jest.config.js
@@ -79,7 +79,7 @@ module.exports = {
     'src/store/reducers/liskService.js',
     'src/tests.js',
     'src/utils/api/ledger.js',
-    'src/utils/api/liskService.js',
+    'src/utils/api/lsk/liskService.js',
     'src/utils/api/btc/',
     'src/utils/applyDeviceClass.js',
     'src/utils/ledger.js',
@@ -116,7 +116,7 @@ module.exports = {
     'app/src/utils.js',
     'app/src/hwManager.js',
     'app/src/trezor.js',
-    'src/utils/api/account.js',
+    'src/utils/api/lsk/account.js',
     'src/utils/api/hwWallet.js',
     'src/utils/loading.js',
     'src/store/middlewares/peers.js',
