+++ src/actions/account.test.js
@@ -19,9 +19,9 @@ import {
 } from './account';
 import { secondPassphraseRegisteredFailure } from './secondPassphrase';
 import { delegateRegisteredFailure } from './delegate';
-import * as accountApi from '../utils/api/account';
+import * as accountApi from '../utils/api/lsk/account';
 import * as delegateApi from '../utils/api/delegate';
-import * as transactionsApi from '../utils/api/transactions';
+import * as transactionsApi from '../utils/api/lsk/transactions';
 import * as blocksApi from '../utils/api/blocks';
 import Fees from '../constants/fees';
 import transactionTypes from '../constants/transactionTypes';
