+++ src/actions/search.js
@@ -1,8 +1,8 @@
 import actionTypes from '../constants/actions';
 import { loadingStarted, loadingFinished } from '../actions/loading';
-import { getAccount } from '../utils/api/account';
+import { getAccount } from '../utils/api/lsk/account';
 import { getDelegate, getVoters, getVotes, listDelegates } from '../utils/api/delegate';
-import { getTransactions } from '../utils/api/transactions';
+import { getTransactions } from '../utils/api/lsk/transactions';
 import { getBlocks } from '../utils/api/blocks';
 import searchAll from '../utils/api/search';
 import transactionTypes from '../constants/transactionTypes';
