+++ src/actions/search.test.js
@@ -2,9 +2,9 @@ import actionTypes from '../constants/actions';
 import networks from '../constants/networks';
 import txFilters from '../constants/transactionFilters';
 import * as searchAPI from '../utils/api/search';
-import * as accountAPI from '../utils/api/account';
+import * as accountAPI from '../utils/api/lsk/account';
 import * as delegateAPI from '../utils/api/delegate';
-import * as transactionsAPI from '../utils/api/transactions';
+import * as transactionsAPI from '../utils/api/lsk/transactions';
 import accounts from '../../test/constants/accounts';
 import * as actions from './search';
 
@@ -18,9 +18,9 @@ const {
 } = actions;
 
 jest.mock('../utils/api/search');
-jest.mock('../utils/api/account');
+jest.mock('../utils/api/lsk/account');
 jest.mock('../utils/api/delegate');
-jest.mock('../utils/api/transactions');
+jest.mock('../utils/api/lsk/transactions');
 
 describe('actions: search', () => {
   let dispatch;
