+++ src/store/middlewares/account.test.js
@@ -5,8 +5,8 @@ import * as transactionsActions from '../../actions/transactions';
 import accountConfig from '../../constants/account';
 import * as votingActions from '../../actions/voting';
 import * as peersActions from '../../actions/peers';
-import * as accountApi from '../../utils/api/account';
-import * as transactionsApi from '../../utils/api/transactions';
+import * as accountApi from '../../utils/api/lsk/account';
+import * as transactionsApi from '../../utils/api/lsk/transactions';
 import * as accountUtils from '../../utils/login';
 import accounts from '../../../test/constants/accounts';
 import networks from '../../constants/networks';
